refactor(model): canonicalize all downloadModels callers to registry/download
Update 27 files: replace every
  from src.model.downloadModels import X, Y, Z
with the canonical split:
  from src.model.registry import <registry-names>
  from src.model.download  import <download-names>

Delete src/model/downloadModels.py now that no callers remain.
