+++ src/extraArches/fastlinedarken.py
@@ -3,7 +3,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from src.model.downloadModels import weightsDir
+from src.model.registry import weightsDir
 from src.infra.isCudaInit import CudaChecker
 
 checker = CudaChecker()
