+++ src/gmfss/gmfss.py
@@ -3,7 +3,9 @@ import torch
 import logging
 from contextlib import nullcontext
 
-from src.model.downloadModels import downloadModels, weightsDir
+from src.model.registry import weightsDir
+
+from src.model.download import downloadModels
 from src.infra.isCudaInit import CudaChecker
 from src.infra.logAndPrint import logAndPrint
