+++ src/interpolate/rife_directml.py
@@ -5,7 +5,9 @@ import torch.nn.functional as F
 import math
 import numpy as np
 
-from src.model.downloadModels import downloadModels, weightsDir, modelsMap, resolveWeightPath
+from src.model.registry import weightsDir, modelsMap
+
+from src.model.download import downloadModels, resolveWeightPath
 from src.infra.isCudaInit import CudaChecker
 from src.infra.logAndPrint import logAndPrint
