+++ src/segment/animeSegment.py
@@ -4,7 +4,9 @@ import os
 import torch
 import torch.nn.functional as F
 
-from src.model.downloadModels import downloadModels, weightsDir, modelsMap, resolveWeightPath
+from src.model.registry import weightsDir, modelsMap
+
+from src.model.download import downloadModels, resolveWeightPath
 from src.io.ffmpegSettings import BuildBuffer, WriteBuffer
 from concurrent.futures import ThreadPoolExecutor
 from src.infra.progressBarLogic import ProgressBarLogic
